Move tx_eth_bridge code to own module
