+++ apps/src/lib/node/ledger/protocol/transactions/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod ethereum_events;
